Extracted email verification to its own class.
